Remove version from generated files
Including the version in every file causes every file to be modified
any time the generator is run. This creates a large diff that obscures the actual changes.
The version is already known in the generation/swagger.json file.

Add warning that file is generated and should not be modified.