Autogenerate the SDK from the OpenAPI spec (3.0.0) (#102)
* Use nullable reference types and datetimeoffset

* Mark non-required types as nullable

* Add required keyword to required properties

* Mark API method parameters as optional

* Allow SDK to handle No Content (204) responses

Previously, the SDK would throw an exception when trying to parse an empty string.
Now we will return the empty string, if a string response is requested.

If the caller was expecting a deserialized type, an exception will still be raised for no content.

* Remove version from generated files

Including the version in every file causes every file to be modified
any time the generator is run. This creates a large diff that obscures the actual changes.
The version is already known in the generation/swagger.json file.

Add warning that file is generated and should not be modified.

* Verify can deserialize create label response

Convert doubles to decimals

* Add optional error properties

Allow arbitrary values in advanced_options.

* added A4 and A6 to label layout enum

* Mark v2 methods as obsolete

* Add generated documentation

* Update supported .NET versions

---------

Co-authored-by: Joshua Flanagan <joshuaflanagan@gmail.com>
Co-authored-by: Matt Burton <matt.burton@gmail.com>
Co-authored-by: Joe Pill <joe.pill@auctane.com>
Co-authored-by: Bibhusan-Shah <bibhusan.shah@auctane.com>
Co-authored-by: Bibhusan-Shah <87672403+Bibhusan-Shah@users.noreply.github.com>